+++ CHANGELOG.md
@@ -4,6 +4,25 @@ All notable changes to this project will be documented in this file.
 See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 
+# 2.1.5 (1st July 2025)
+
+### Added
+
+Feature: Support Marketplace as a payment method for Essentials subscription
+Feature: Add TLS certificate to databases’ data sources
+
+### Fixed:
+
+Unexpected state `dynamic-endpoints-creation-pending'
+Can not disable default user on essentials db
+
+# 2.1.4 (22nd May 2025)
+
+### Added
+
+- Documentation for `rediscloud_active_active_subscription_regions` added.
+- Schema documentation amended to match documentation above.
+
 # 2.1.3 (21st May 2025)
 
 ### Added